feat: initialize configura early
