+++ yui-common/src/main/java/me/whereareiam/yui/common/CommonConfiguration.java
@@ -1,5 +1,6 @@
 package me.whereareiam.yui.common;
 
+import jakarta.annotation.PreDestroy;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import me.whereareiam.yui.model.config.settings.Settings;
@@ -7,6 +8,7 @@ import net.dv8tion.jda.api.JDA;
 import net.dv8tion.jda.api.JDABuilder;
 import net.dv8tion.jda.api.utils.ChunkingFilter;
 import net.dv8tion.jda.api.utils.MemberCachePolicy;
+import org.springframework.beans.factory.ObjectProvider;
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
 import org.springframework.context.annotation.Primary;
@@ -18,9 +20,12 @@ import java.util.concurrent.Executors;
 @Configuration
 @RequiredArgsConstructor
 public class CommonConfiguration {
+	private final ObjectProvider<Settings> settingsProvider;
+
 	@Bean
 	@Primary
-	public JDA jda(Settings settings) {
+	public JDA jda() {
+		Settings settings = settingsProvider.getObject();
 		JDA jda;
 
 		try {
@@ -38,20 +43,19 @@ public class CommonConfiguration {
 
 			jda = builder.build().awaitReady();
 
-			jda.getGuilds()
-					.stream()
-					.filter(guild -> !guild.getId().equals(settings.getDiscord().getGuildId()))
-					.forEach(guild -> {
-						log.info("Leaving guild '{}' ({})", guild.getName(), guild.getId());
-						guild.leave().queue();
-					});
-
-
 		} catch (InterruptedException e) {
 			throw new RuntimeException(e);
 		}
 
 		return jda;
+	}
+
+	@PreDestroy
+	public void shutdownJda(JDA jda) {
+		if (jda == null) return;
+
+		log.info("Shutting down YUI");
+		jda.shutdown();
 	}
 
 	@Bean(destroyMethod = "shutdown")
