+++ yui-common/src/main/java/me/whereareiam/yui/common/config/ConfiguraBootstrap.java
@@ -6,37 +6,23 @@ import me.whereareiam.configura.Config;
 import me.whereareiam.configura.reader.ConfigReader;
 import me.whereareiam.configura.type.Format;
 import me.whereareiam.configura.writer.ConfigWriter;
-import me.whereareiam.yui.LifecycleTask;
 import me.whereareiam.yui.common.config.adapter.ColorAdapter;
 import me.whereareiam.yui.common.config.adapter.DiscordLocaleAdapter;
 import me.whereareiam.yui.config.ConfigurationTypeResolver;
-import me.whereareiam.yui.registry.Registry;
 import me.whereareiam.yui.type.ConfigurationType;
 import net.dv8tion.jda.api.interactions.DiscordLocale;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Component;
 
 import java.awt.*;
-import java.util.concurrent.CompletableFuture;
 
 @Component
 @RequiredArgsConstructor(onConstructor_ = @Autowired)
-public class ConfiguraBootstrap implements LifecycleTask {
+public class ConfiguraBootstrap {
 	private final ConfigurationTypeResolver resolver;
-	private final Registry<LifecycleTask> lifecycleRegistry;
 
 	@PostConstruct
-	public void registerSelf() {
-		lifecycleRegistry.register(this);
-	}
-
-	@Override
-	public String getName() {
-		return "BOOTSTRAP_CONFIGURA";
-	}
-
-	@Override
-	public CompletableFuture<Void> start() {
+	public void initConfigura() {
 		// Resolve the preferred configuration format
 		ConfigurationType type = resolver.getConfigurationType();
 		Format format = (type == ConfigurationType.JSON) ? Format.JSON : Format.YAML;
@@ -50,8 +36,5 @@ public class ConfiguraBootstrap implements LifecycleTask {
 		// Register adapters
 		Config.registerAdapter(Color.class, ColorAdapter.class);
 		Config.registerAdapter(DiscordLocale.class, DiscordLocaleAdapter.class);
-
-		return CompletableFuture.completedFuture(null);
 	}
 }
-
