+++ yui-common/src/main/java/me/whereareiam/yui/common/config/provider/style/EmbedsProvider.java
@@ -1,11 +1,9 @@
 package me.whereareiam.yui.common.config.provider.style;
 
 import me.whereareiam.configura.Config;
-import me.whereareiam.yui.Reloadable;
 import me.whereareiam.yui.common.config.provider.DefaultConfigProvider;
 import me.whereareiam.yui.common.config.template.style.EmbedsTemplate;
 import me.whereareiam.yui.model.config.style.embed.EmbedStyle;
-import me.whereareiam.yui.registry.Registry;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.beans.factory.annotation.Qualifier;
 import org.springframework.stereotype.Component;
@@ -15,11 +13,12 @@ import java.nio.file.Path;
 @Component
 public class EmbedsProvider extends DefaultConfigProvider<EmbedStyle> {
 	@Autowired
-	public EmbedsProvider(
-			@Qualifier("stylesPath") Path stylesPath,
-			Registry<Reloadable> registry
-	) {
-		super(stylesPath, registry);
+	@Qualifier("stylesPath")
+	private Path stylesPath;
+
+	@Override
+	protected Path getBasePath() {
+		return stylesPath;
 	}
 
 	@Override
