+++ yui-common/src/main/java/me/whereareiam/yui/common/initialization/RolesInitializationService.java
@@ -1,4 +1,4 @@
-package me.whereareiam.yui.common.service.initialization;
+package me.whereareiam.yui.common.initialization;
 
 import lombok.AllArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
