+++ yui-common/src/main/java/me/whereareiam/yui/common/initialization/tasks/InitializeTranslationsTask.java
@@ -1,12 +1,11 @@
-package me.whereareiam.yui.common.service.initialization.tasks;
+package me.whereareiam.yui.common.initialization.tasks;
 
 import jakarta.annotation.PostConstruct;
 import lombok.RequiredArgsConstructor;
 import me.whereareiam.yui.registry.Registry;
 import me.whereareiam.yui.LifecycleTask;
-import me.whereareiam.yui.common.scanner.ComponentListenerScanner;
+import me.whereareiam.yui.common.translation.DefaultTranslationService;
 import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.context.ApplicationContext;
 import org.springframework.stereotype.Component;
 
 import java.util.List;
@@ -14,9 +13,8 @@ import java.util.concurrent.CompletableFuture;
 
 @Component
 @RequiredArgsConstructor(onConstructor_ = @Autowired)
-public class ScanComponentListenersTask implements LifecycleTask {
-	private final ComponentListenerScanner scanner;
-	private final ApplicationContext ctx;
+public class InitializeTranslationsTask implements LifecycleTask {
+	private final DefaultTranslationService translations;
 	private final Registry<LifecycleTask> lifecycleRegistry;
 
 	@PostConstruct
@@ -26,17 +24,17 @@ public class ScanComponentListenersTask implements LifecycleTask {
 
 	@Override
 	public String getName() {
-		return "SCAN_COMPONENT_LISTENERS";
+		return "INIT_TRANSLATIONS";
 	}
 
 	@Override
 	public List<String> getDependencies() {
-		return List.of("INIT_PLUGINS");
+		return List.of("PURGE_TEMP_CHANNELS");
 	}
 
 	@Override
 	public CompletableFuture<Void> start() {
-		scanner.scan(ctx);
+		translations.initialize();
 		return CompletableFuture.completedFuture(null);
 	}
 }
