+++ yui-common/src/main/java/me/whereareiam/yui/common/initialization/tasks/ScanComponentListenersTask.java
@@ -1,11 +1,12 @@
-package me.whereareiam.yui.common.service.initialization.tasks;
+package me.whereareiam.yui.common.initialization.tasks;
 
 import jakarta.annotation.PostConstruct;
 import lombok.RequiredArgsConstructor;
 import me.whereareiam.yui.registry.Registry;
 import me.whereareiam.yui.LifecycleTask;
-import me.whereareiam.yui.common.translation.DefaultTranslationService;
+import me.whereareiam.yui.common.scanner.ComponentListenerScanner;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.context.ApplicationContext;
 import org.springframework.stereotype.Component;
 
 import java.util.List;
@@ -13,8 +14,9 @@ import java.util.concurrent.CompletableFuture;
 
 @Component
 @RequiredArgsConstructor(onConstructor_ = @Autowired)
-public class InitializeTranslationsTask implements LifecycleTask {
-	private final DefaultTranslationService translations;
+public class ScanComponentListenersTask implements LifecycleTask {
+	private final ComponentListenerScanner scanner;
+	private final ApplicationContext ctx;
 	private final Registry<LifecycleTask> lifecycleRegistry;
 
 	@PostConstruct
@@ -24,17 +26,17 @@ public class InitializeTranslationsTask implements LifecycleTask {
 
 	@Override
 	public String getName() {
-		return "INIT_TRANSLATIONS";
+		return "SCAN_COMPONENT_LISTENERS";
 	}
 
 	@Override
 	public List<String> getDependencies() {
-		return List.of("PURGE_TEMP_CHANNELS");
+		return List.of("INIT_PLUGINS");
 	}
 
 	@Override
 	public CompletableFuture<Void> start() {
-		translations.initialize();
+		scanner.scan(ctx);
 		return CompletableFuture.completedFuture(null);
 	}
 }
