+++ yui-common/src/main/java/me/whereareiam/yui/common/listener/GuildMemberJoinListener.java
@@ -2,7 +2,7 @@ package me.whereareiam.yui.common.listener;
 
 import lombok.AllArgsConstructor;
 import me.whereareiam.yui.service.UserRoleService;
-import me.whereareiam.yui.common.service.initialization.UserProfileInitializationService;
+import me.whereareiam.yui.common.initialization.UserProfileInitializationService;
 import net.dv8tion.jda.api.events.guild.member.GuildMemberJoinEvent;
 import net.dv8tion.jda.api.hooks.ListenerAdapter;
 import org.springframework.stereotype.Component;
